server-1/requirements.txt
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
pymongo==4.6.1